refactor(util): move bundle staleness reporting into its own module
Second and last cut of #293's three concerns. `dependencies.py` goes
2033 -> 1770 lines and is now asset access plus CDN version resolution,
which is what the issue asked for.

`_RELEASE_RE` stays behind again, for the reason recorded on #293 after
the first cut: it sits inside the moved section but `_version_key` uses
it, and it is a version-parsing primitive rather than a freshness one.

The edge back into `dependencies` is one call -- `warn_bundle_unreadable`
from `inline_bundle_tags` -- imported inside the function so the
module-scope dependency stays one-way.

Reads that tests patch go through the module object rather than a
from-import: `_deps.maidr_js_version()` and `_deps._published_version()`.
A from-import binds the object at import time, so patching
`dependencies.maidr_js_version` would have silently stopped reaching this
module, and 29 tests said so.

Refs #293
