+++ tests/core/test_resolver_outcome.py
@@ -25,6 +25,7 @@ from urllib.error import HTTPError
 
 import pytest
 
+from maidr.util import bundle_freshness as freshness
 from maidr.util import dependencies
 
 
@@ -70,7 +71,7 @@ def test_nothing_is_reported_before_a_lookup_runs(clean) -> None:
     A pinned or offline session never resolves. Reading an empty outcome as
     a verdict there would report every endpoint as fine, on no evidence.
     """
-    assert dependencies.resolver_outcome() is None
+    assert freshness.resolver_outcome() is None
 
 
 def test_an_outage_is_reported_as_unreachable(monkeypatch, clean) -> None:
@@ -82,7 +83,7 @@ def test_an_outage_is_reported_as_unreachable(monkeypatch, clean) -> None:
 
     assert dependencies._fetch_latest_version(3.0) is None
 
-    outcome = dependencies.resolver_outcome()
+    outcome = freshness.resolver_outcome()
     assert len(outcome.unreachable) == ENDPOINT_COUNT
     assert outcome.answered_badly == ()
 
@@ -101,7 +102,7 @@ def test_an_http_error_is_an_answer(monkeypatch, clean) -> None:
 
     assert dependencies._fetch_latest_version(3.0) is None
 
-    outcome = dependencies.resolver_outcome()
+    outcome = freshness.resolver_outcome()
     assert len(outcome.answered_badly) == ENDPOINT_COUNT
     assert outcome.unreachable == ()
 
@@ -116,7 +117,7 @@ def test_a_payload_without_the_key_is_an_answer(monkeypatch, clean) -> None:
     answer_with(monkeypatch, json_body({"something_else": "4.2.0"}))
 
     assert dependencies._fetch_latest_version(3.0) is None
-    assert len(dependencies.resolver_outcome().answered_badly) == ENDPOINT_COUNT
+    assert len(freshness.resolver_outcome().answered_badly) == ENDPOINT_COUNT
 
 
 def test_a_value_that_is_not_a_version_is_an_answer(monkeypatch, clean) -> None:
@@ -129,7 +130,7 @@ def test_a_value_that_is_not_a_version_is_an_answer(monkeypatch, clean) -> None:
     )
 
     assert dependencies._fetch_latest_version(3.0) is None
-    assert len(dependencies.resolver_outcome().answered_badly) == ENDPOINT_COUNT
+    assert len(freshness.resolver_outcome().answered_badly) == ENDPOINT_COUNT
 
 
 def test_a_body_that_will_not_parse_is_an_answer(monkeypatch, clean) -> None:
@@ -140,7 +141,7 @@ def test_a_body_that_will_not_parse_is_an_answer(monkeypatch, clean) -> None:
     )
 
     assert dependencies._fetch_latest_version(3.0) is None
-    assert len(dependencies.resolver_outcome().answered_badly) == ENDPOINT_COUNT
+    assert len(freshness.resolver_outcome().answered_badly) == ENDPOINT_COUNT
 
 
 def test_an_unrecognised_failure_counts_as_unreachable(monkeypatch, clean) -> None:
@@ -161,7 +162,7 @@ def test_an_unrecognised_failure_counts_as_unreachable(monkeypatch, clean) -> No
 
     assert dependencies._fetch_latest_version(3.0) is None
 
-    outcome = dependencies.resolver_outcome()
+    outcome = freshness.resolver_outcome()
     assert len(outcome.unreachable) == ENDPOINT_COUNT
     assert outcome.answered_badly == ()
 
@@ -176,7 +177,7 @@ def test_a_success_reports_no_failures(monkeypatch, clean) -> None:
 
     assert dependencies._fetch_latest_version(3.0) == "4.2.0"
 
-    outcome = dependencies.resolver_outcome()
+    outcome = freshness.resolver_outcome()
     assert outcome == dependencies.ResolverOutcome("4.2.0", (), ())
 
 
@@ -201,7 +202,7 @@ def test_a_fallback_records_the_endpoint_it_fell_back_from(
 
     assert dependencies._fetch_latest_version(3.0) == "9.9.9"
 
-    outcome = dependencies.resolver_outcome()
+    outcome = freshness.resolver_outcome()
     assert outcome.resolved == "9.9.9"
     assert len(outcome.unreachable) == 1
     assert outcome.answered_badly == ()
@@ -220,11 +221,11 @@ def test_a_reset_discards_the_verdict_with_the_lookup(monkeypatch, clean) -> Non
 
     answer_with(monkeypatch, not_found)
     dependencies._fetch_latest_version(3.0)
-    assert dependencies.resolver_outcome().answered_badly
+    assert freshness.resolver_outcome().answered_badly
 
     dependencies.reset_cdn_version_cache()
 
-    assert dependencies.resolver_outcome() is None
+    assert freshness.resolver_outcome() is None
 
 
 def test_the_render_path_is_unchanged_by_any_of_it(monkeypatch, clean) -> None:
@@ -240,4 +241,4 @@ def test_the_render_path_is_unchanged_by_any_of_it(monkeypatch, clean) -> None:
     )
 
     assert dependencies.get_cdn_version() == dependencies.maidr_js_version()
-    assert dependencies.bundle_status(resolve=False).published is None
+    assert freshness.bundle_status(resolve=False).published is None
